[Feat] 시뮬레이션 날짜 변경시 매수예정내역 삭제, 주식 거래내역 등록 구현
시뮬레이션 날짜가 변경되면 예정 매수 테이블과 현재주식가격과 비교하여 구매할 수 있다면 회원 거래내역에 추가 하고 매수 예정리스트에 해당 종목 제거함

Issue:S08P12B111-528, S08P12B111-521

+++ back/src/main/java/com/udteam/miristock/dto/StockDealDto.java
@@ -20,7 +20,7 @@ public class StockDealDto {
     private Integer stockDealDate;
     private Integer stockDealBuyClosingPrice;
     private Integer stockDealSellClosingPrice;
-    private Integer stockDealAmount;
+    private Long stockDealAmount;
     private Deal stockDealType;
     private Float stockDealEarnRate;
     private Long stockDealEarnPrice;
